fix(openui-dashboard): remove server-side tool calls and add simulated tool latency
Remove tools from the runTools() call so the LLM only generates Query()
statements instead of executing tools server-side during generation.

Replace SSE streaming helper (sseResponseFromRunner) with a manual
ReadableStream that pipes plain create() chunks directly as SSE events
with an explicit data: [DONE] terminator.

Add NEXT_PUBLIC_TOOL_DELAY_MS env var so mock tool execute() functions
simulate network latency, surfacing intermediate loading states in the UI.
