+++ examples/openui-dashboard/.env.example
@@ -1,0 +1,1 @@
+NEXT_PUBLIC_TOOL_DELAY_MS=
